+++ test/lib/backends/backend.test.js
@@ -3,7 +3,7 @@ const assert = require('assert')
 
 describe('Backend super class', () => {
   it('should export necessary functions', () => {
-    const backend = new Backend({ foo: 'bar' })
+    const backend = new Backend({ config: { foo: 'bar' } })
     assert.strictEqual(typeof backend.setup, 'function')
     assert.strictEqual(typeof backend.clean, 'function')
     assert.strictEqual(typeof backend.isHealthy, 'function')
@@ -14,22 +14,22 @@ describe('Backend super class', () => {
   })
 
   it('should store config in itself', () => {
-    const backend = new Backend({ foo: 'bar' })
+    const backend = new Backend({ config: { foo: 'bar' } })
     assert.deepStrictEqual(backend.config, { foo: 'bar' })
   })
 
   it('should call cb for setup', (done) => {
-    const backend = new Backend({ foo: 'bar' })
+    const backend = new Backend({ config: { foo: 'bar' } })
     backend.setup(done)
   })
 
   it('should call cb for clean', (done) => {
-    const backend = new Backend({ foo: 'bar' })
+    const backend = new Backend({ config: { foo: 'bar' } })
     backend.clean(done)
   })
 
   it('should call cb for isHealthy', (done) => {
-    const backend = new Backend({ foo: 'bar' })
+    const backend = new Backend({ config: { foo: 'bar' } })
     backend.isHealthy((error, healthy) => {
       assert.strictEqual(error, null)
       // `false` because this function _must_ be implemented
@@ -39,7 +39,7 @@ describe('Backend super class', () => {
   })
 
   it('should return nothing for create* streams', () => {
-    const backend = new Backend({ foo: 'bar' })
+    const backend = new Backend({ config: { foo: 'bar' } })
     assert.strictEqual(typeof backend.createDataReadStream(), 'undefined')
     assert.strictEqual(typeof backend.createCacheReadStream(), 'undefined')
     assert.strictEqual(typeof backend.createDataWriteStream(), 'undefined')